Move totp into login.rs
