+++ crates/bitwarden-exporters/src/cxf/login.rs
@@ -6,18 +6,42 @@
 use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine};
 use bitwarden_core::MissingFieldError;
 use bitwarden_fido::{string_to_guid_bytes, InvalidGuid};
-use bitwarden_vault::FieldType;
+use bitwarden_vault::{FieldType, Totp, TotpAlgorithm};
 use chrono::{DateTime, Utc};
 use credential_exchange_format::{
-    AndroidAppIdCredential, BasicAuthCredential, CredentialScope, PasskeyCredential, TotpCredential,
+    AndroidAppIdCredential, BasicAuthCredential, CredentialScope, OTPHashAlgorithm,
+    PasskeyCredential, TotpCredential,
 };
 use thiserror::Error;
 
-use super::totp::totp_credential_to_totp;
 use crate::{Fido2Credential, Field, Login, LoginUri};
 
 /// Prefix that indicates the URL is an Android app scheme.
 const ANDROID_APP_SCHEME: &str = "androidapp://";
+
+/// Convert CXF TotpCredential to Bitwarden's Totp struct
+/// This ensures we use the exact same encoding and formatting as Bitwarden's core implementation
+fn totp_credential_to_totp(cxf_totp: &TotpCredential) -> Totp {
+    let algorithm = match cxf_totp.algorithm {
+        OTPHashAlgorithm::Sha1 => TotpAlgorithm::Sha1,
+        OTPHashAlgorithm::Sha256 => TotpAlgorithm::Sha256,
+        OTPHashAlgorithm::Sha512 => TotpAlgorithm::Sha512,
+        OTPHashAlgorithm::Unknown(ref algo) if algo == "steam" => TotpAlgorithm::Steam,
+        OTPHashAlgorithm::Unknown(_) | _ => TotpAlgorithm::Sha1, /* Default to SHA1 for unknown
+                                                                  * algorithms */
+    };
+
+    let secret_bytes: Vec<u8> = cxf_totp.secret.clone().into();
+
+    Totp {
+        account: cxf_totp.username.clone(),
+        algorithm,
+        digits: cxf_totp.digits as u32,
+        issuer: cxf_totp.issuer.clone(),
+        period: cxf_totp.period as u32,
+        secret: secret_bytes,
+    }
+}
 
 pub(super) fn to_login(
     creation_date: DateTime<Utc>,
@@ -474,5 +498,49 @@ mod tests {
                 },
             ]
         );
+    }
+
+    // TOTP tests
+    #[test]
+    fn test_totp_credential_to_totp_basic() {
+        let totp = TotpCredential {
+            secret: "Hello World!".as_bytes().to_vec().into(),
+            period: 30,
+            digits: 6,
+            username: Some("test@example.com".to_string()),
+            algorithm: OTPHashAlgorithm::Sha1,
+            issuer: Some("Example".to_string()),
+        };
+
+        let bitwarden_totp = totp_credential_to_totp(&totp);
+        let otpauth = bitwarden_totp.to_string();
+
+        assert!(otpauth.starts_with("otpauth://totp/Example:test%40example%2Ecom?secret="));
+        assert!(otpauth.contains("&issuer=Example"));
+        // Default period (30) and digits (6) and algorithm (SHA1) should not be included
+        assert!(!otpauth.contains("&period=30"));
+        assert!(!otpauth.contains("&digits=6"));
+        assert!(!otpauth.contains("&algorithm=SHA1"));
+    }
+
+    #[test]
+    fn test_totp_credential_to_totp_custom_parameters() {
+        let totp = TotpCredential {
+            secret: "Hello World!".as_bytes().to_vec().into(),
+            period: 60,
+            digits: 8,
+            username: Some("user".to_string()),
+            algorithm: OTPHashAlgorithm::Sha256,
+            issuer: Some("Custom Issuer".to_string()),
+        };
+
+        let bitwarden_totp = totp_credential_to_totp(&totp);
+        let otpauth = bitwarden_totp.to_string();
+
+        assert!(otpauth.contains("Custom%20Issuer:user"));
+        assert!(otpauth.contains("&issuer=Custom%20Issuer"));
+        assert!(otpauth.contains("&period=60"));
+        assert!(otpauth.contains("&digits=8"));
+        assert!(otpauth.contains("&algorithm=SHA256"));
     }
 }
