+++ crates/bitwarden-exporters/src/cxf/mod.rs
@@ -18,5 +18,4 @@ mod editable_field;
 mod identity;
 mod login;
 mod note;
-mod totp;
 mod wifi;
